requirements.txt
flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.2
gunicorn==21.2.0
jupyter==1.0.0